fix(kwavers): restore CartesianPosition and fix kwavers-python typed-quantity boundary
- Add CartesianPosition to kwavers-transducer::transducers::physics (position.rs)
  A 3-D SI position type wrapping [Length<f64>; 3] with from_base([f64;3]) and
  new([Length<f64>;3]) constructors returning KwaversResult<Self>. Includes
  into_base() (aequitas convention), x/y/z accessors, distance_to(), and
  coordinate-range validation (finite, ±10 m). Resolves kwavers-therapy compile
  failure: CartesianPosition was removed from transducers/physics but hifu_planning
  types still used it.

- Fix kwavers-python lens.rs PyO3 typed-quantity boundary
  materials.rs functions were typed with aequitas Length/Velocity/Frequency/Time
  quantities (correct SSOT). lens.rs PyO3 wrappers still used raw f64. Updated
  all four functions to convert at the PyO3 boundary:
    - acoustic_lens_delay_profile: [f64] → Vec<Length>, f64 → Velocity, Vec<Time> → Vec<f64>
    - fresnel_zone_radii: f64 → Length (x3), Vec<Length> → Vec<f64>
    - isoplanatic_steering_curve: f64 → Length (x2), (Angle,Length) → (f64,f64)
    - corrective_lens_thickness: f64 → Frequency/Velocity/Length, Vec<Length> → Vec<f64>

+++ crates/kwavers-physics/src/acoustics/wave_propagation/nonlinear/burgers/solution.rs
@@ -35,7 +35,7 @@ pub fn fubini_harmonic_amplitude(n: u32, sigma: f64) -> f64 {
     } else if sigma < 1.0 {
         // Pre-shock: Fubini Bessel series.
         let arg = n as f64 * sigma;
-        2.0 * bessel_j(n, arg) / arg
+        2.0 * bessel_j(n as usize, arg) / arg
     } else {
         // Post-shock: Fay sawtooth, B_n = 2/(n(1+sigma)) (Blackstock 1966);
         // matches analytical::wave::sawtooth_harmonic_amplitude.
