+++ crates/kwavers-physics/src/analytical/wave/nonlinear.rs
@@ -1,7 +1,7 @@
 use apollo::fft_1d_leto;
 use kwavers_core::constants::numerical::TWO_PI;
 use kwavers_math::signal::window::hann;
-use kwavers_math::special::bessel::jn;
+use leto_ops::application::special::jn;
 
 /// Evaluate the normalised amplitude of the nth harmonic at nonlinear parameter σ.
 ///
@@ -25,7 +25,7 @@ pub fn fubini_harmonic_amplitude(n: u32, sigma: f64) -> f64 {
     if x.abs() < 1e-15 {
         return if n == 1 { 1.0 } else { 0.0 };
     }
-    2.0 / x * jn(n, x)
+    2.0 / x * jn(n as usize, x)
 }
 
 /// Compute the Fubini harmonic spectrum for harmonics n = 1..=n_max at parameter σ.
